backend/requirements.txt
transformers
accelerate
bitsandbytes
torch
youtube-transcript-api
chromadb
fastapi
uvicorn
requests
python-dotenv
yt-dlp
